% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Euclidean distance between two keypoints}
\usage{
distance(dat, a, b, time)
}
\arguments{
\item{dat}{Reshaped keypoint data (see \code{\link[=reshaping]{reshaping()}}).}

\item{a, b}{Keypoint indices or names.}

\item{time}{Frame index or vector of frames.}
}
\value{
The distance(s) between keypoints \code{a} and \code{b}.
}
\description{
Euclidean distance between two keypoints
}
\examples{
## a minimal reshaped pose: 18 keypoints over three frames
parts <- c("nose","neck","Rshoulder","Relbow","Rwrist","Lshoulder","Lelbow",
  "Lwrist","Rhip","Rknee","Rankle","Lhip","Lknee","Lankle","Reye","Leye",
  "Rear","Lear")
bx <- c(0,0,-.25,-.3,-.32,.25,.3,.32,-.18,-.2,-.22,.18,.2,.22,-.05,.05,-.1,.1)
by <- c(-.3,0,.05,.5,.95,.05,.5,.95,1,1.65,2.28,1,1.65,2.28,-.35,-.35,-.3,-.3)
dat <- list(x = matrix(bx, 3, 18, byrow = TRUE, dimnames = list(NULL, parts)),
            y = matrix(by, 3, 18, byrow = TRUE, dimnames = list(NULL, parts)))
## shoulder width over frames 1:3 (by name) and hip width (by index)
distance(dat, "Rshoulder", "Lshoulder", 1:3)
distance(dat, 9, 12, 1)
}
