% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdegree.R
\name{hdegree}
\alias{hdegree}
\title{Hip joint angle between three keypoints}
\usage{
hdegree(dat, a, b, c, time)
}
\arguments{
\item{dat}{Reshaped keypoint data (see \code{\link[=reshaping]{reshaping()}}).}

\item{a, b, c}{Keypoint indices or names (angle measured at \code{b}).}

\item{time}{Frame index or vector of frames.}
}
\value{
The hip angle(s) in degrees.
}
\description{
Hip joint angle between three keypoints
}
\examples{
## a minimal reshaped pose (one frame is enough for a single angle)
parts <- c("nose","neck","Rshoulder","Relbow","Rwrist","Lshoulder","Lelbow",
  "Lwrist","Rhip","Rknee","Rankle","Lhip","Lknee","Lankle","Reye","Leye",
  "Rear","Lear")
bx <- c(0,0,-.25,-.3,-.32,.25,.3,.32,-.18,-.2,-.22,.18,.2,.22,-.05,.05,-.1,.1)
by <- c(-.3,0,.05,.5,.95,.05,.5,.95,1,1.65,2.28,1,1.65,2.28,-.35,-.35,-.3,-.3)
dat <- list(x = matrix(bx, 2, 18, byrow = TRUE, dimnames = list(NULL, parts)),
            y = matrix(by, 2, 18, byrow = TRUE, dimnames = list(NULL, parts)))
## right hip flexion at the hip (neck - Rhip - Rknee) in frame 1
hdegree(dat, 2, 9, 10, 1)
}
