% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_segment.R
\name{plt.segment}
\alias{plt.segment}
\title{Create segment line for frame plot}
\usage{
plt.segment(x, y)
}
\arguments{
\item{x}{Numeric vector of keypoint x-coordinates for one frame.}

\item{y}{Numeric vector of keypoint y-coordinates for one frame.}
}
\value{
Called for its side effect of drawing the segments; returns \code{NULL}.
}
\description{
Create segment line for frame plot
}
\examples{
bx <- c(0,0,-.25,-.3,-.32,.25,.3,.32,-.18,-.2,-.22,.18,.2,.22,-.05,.05,-.1,.1)
by <- c(-.3,0,.05,.5,.95,.05,.5,.95,1,1.65,2.28,1,1.65,2.28,-.35,-.35,-.3,-.3)
## open a plot of the keypoints first, then draw the skeleton segments
plot(bx, by, ylim = c(2.5, -0.5), asp = 1, pch = 19, main = "one pose")
plt.segment(bx, by)
}
