% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_frames.R
\name{diplay_frames}
\alias{diplay_frames}
\title{Title Plotting frames of gait cycle}
\usage{
diplay_frames(obj, layout = c(5, 5), xlim = c(450, 800), ylim = c(700, 200))
}
\arguments{
\item{obj}{list for reshaped object}

\item{layout}{vector indicating layout for display}

\item{xlim}{vector of 2 elements}

\item{ylim}{vector of 2 elements}
}
\value{
null
}
\description{
Title Plotting frames of gait cycle
}
