% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json.R
\name{import_json}
\alias{import_json}
\title{Importing json files in a batch}
\usage{
import_json(
  myfiles,
  parts = c("nose", "neck", "Rshoulder", "Relbow", "Rwrist", "Lshoulder", "Lelbow",
    "Lwrist", "Rhip", "Rknee", "Rankle", "Lhip", "Lknee", "Lankle", "Reye", "Leye",
    "Rear", "Lear")
)
}
\arguments{
\item{myfiles}{vector of filenames.}

\item{parts}{}
}
\value{

}
\description{
Importing json files in a batch
}
