% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_segment.R
\name{plt.segment}
\alias{plt.segment}
\title{Create segment line for frame plot}
\usage{
plt.segment(x, y)
}
\arguments{
\item{x}{scalar of x-coordinate}

\item{y}{scalar of y-coordinate}
}
\description{
internal function
}
